requirements.txt
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
scipy>=1.11
yfinance>=0.2.36
matplotlib>=3.7
pyarrow>=14.0
pytest>=7.4
pyyaml>=6.0
